iconos: ...

+++ scripts/svg_converter.sh
@@ -4,8 +4,8 @@
 svgs='resources/svg'
 images='resources/images'
 
-color_a="rgb(200,200,200)"
-color_b="rgb(150,150,150)"
+color_a="rgb(150,150,150)"
+color_b="rgb(190,190,190)"
 color_c="#ff5e00"
 
 size=128
@@ -21,22 +21,24 @@ function export_png() {
     svg_temp="$images/$filename"_temp_.svg
     output="$images/$filename"_"$suffix".png
 
-    cp $svg $svg_temp
+    if [ ! -f $output ]; then
+        cp $svg $svg_temp
 
-    # cambia el color 'fill'
-    svg_temp
-    sed -i "s|fill=\"none\"|fill=\"$color\"|g" $svg_temp
-    #
+        # cambia el color 'fill'
+        svg_temp
+        sed -i "s|fill=\"none\"|fill=\"$color\"|g" $svg_temp
+        #
 
-    # cambia tamaño
-    sed -i "s|height=\"24\"|height=\"$size\"|g" $svg_temp
-    sed -i "s|width=\"24\"|width=\"$size\"|g" $svg_temp
-    #
-    #
+        # cambia tamaño
+        sed -i "s|height=\"24\"|height=\"$size\"|g" $svg_temp
+        sed -i "s|width=\"24\"|width=\"$size\"|g" $svg_temp
+        #
+        #
 
-    inkscape $svg_temp -e $output
+        inkscape $svg_temp -e $output
 
-    rm $svg_temp
+        rm $svg_temp
+    fi
 
 }
